README: move supported hardware features in a table
Without that change we don't really know on which Raspberry Pi
hardware features have been confirmed to work.

While it's pretty clear that only the Raspberry Pi 2 is well tested,
it's better to show empty entries for the other Raspberry Pi when
unsure as not only it is more clear, but it could also push people to
test the known working features on other Raspberry Pi and then to
send patches to update the README.

I followed the specifications[1] to do the table conversion.

[1]https://github.github.com/gfm/#tables-extension-

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>